+++ CHANGELOG.md
@@ -1,6 +1,12 @@
 # 1.5.22 Release:
 ## Update Notes:
 - This is not a V14 update. An update for V14 is still a work in progress.
+## New Legacy Content:
+- Cruel
+
+# 1.5.22 Release:
+## Update Notes:
+- This is not a V14 update. An update for V14 is still a work in progress.
 ## New Modern Content
 - Death Armor
 - Elemental Attunement